# server/api/openapi/phannd.me-api-documentation.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "PhanND - phannd.me API Documentation API",
    "description": "The phannd.me API Documentation API description",
    "version": "1.0",
    "contact": {}
  },
  "servers": [],
  "security": [
    {
      "apiKeyCookie": [],
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/posts": {
      "get": {
        "operationId": "PostController_getPublishedPosts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Số trang hiện tại",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số lượng bản ghi trên mỗi trang",
            "schema": {
              "minimum": 1,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Từ khóa tìm kiếm theo tiêu đề hoặc nội dung",
            "schema": {
              "example": "nestjs",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Trường dữ liệu để sắp xếp",
            "schema": {
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Thứ tự sắp xếp",
            "schema": {
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "authorId",
            "required": false,
            "in": "query",
            "description": "Lọc theo ID tác giả",
            "schema": {
              "format": "uuid",
              "example": "d6e9f5d8-6be6-4bfe-8f13-29de6f4f2c11",
              "type": "string"
            }
          },
          {
            "name": "categoryIds",
            "required": false,
            "in": "query",
            "description": "Lọc theo danh sách ID danh mục",
            "schema": {
              "example": [
                "b733b08a-46bd-4e85-bd60-ff9e789417c2"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "published",
            "required": false,
            "in": "query",
            "description": "Lọc theo trạng thái xuất bản",
            "schema": {
              "example": "true",
              "type": "string"
            }
          },
          {
            "name": "createdAtFrom",
            "required": false,
            "in": "query",
            "description": "Ngày tạo bắt đầu để lọc",
            "schema": {
              "format": "date",
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "createdAtTo",
            "required": false,
            "in": "query",
            "description": "Ngày tạo kết thúc để lọc",
            "schema": {
              "format": "date",
              "example": "2026-08-31",
              "type": "string"
            }
          },
          {
            "name": "updatedAtFrom",
            "required": false,
            "in": "query",
            "description": "Ngày cập nhật bắt đầu để lọc",
            "schema": {
              "format": "date",
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "updatedAtTo",
            "required": false,
            "in": "query",
            "description": "Ngày cập nhật kết thúc để lọc",
            "schema": {
              "format": "date",
              "example": "2026-08-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The serialized paginated result of PostBriefResponseDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PostBriefResponseDto"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      },
      "post": {
        "operationId": "PostController_createPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostBaseCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The serialized result of Creation result",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OperationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
    "/api/v1/posts/all": {
      "get": {
        "operationId": "PostController_getAllPosts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Số trang hiện tại",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số lượng bản ghi trên mỗi trang",
            "schema": {
              "minimum": 1,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Từ khóa tìm kiếm theo tiêu đề hoặc nội dung",
            "schema": {
              "example": "nestjs",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Trường dữ liệu để sắp xếp",
            "schema": {
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Thứ tự sắp xếp",
            "schema": {
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "authorId",
            "required": false,
            "in": "query",
            "description": "Lọc theo ID tác giả",
            "schema": {
              "format": "uuid",
              "example": "d6e9f5d8-6be6-4bfe-8f13-29de6f4f2c11",
              "type": "string"
            }
          },
          {
            "name": "categoryIds",
            "required": false,
            "in": "query",
            "description": "Lọc theo danh sách ID danh mục",
            "schema": {
              "example": [
                "b733b08a-46bd-4e85-bd60-ff9e789417c2"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "published",
            "required": false,
            "in": "query",
            "description": "Lọc theo trạng thái xuất bản",
            "schema": {
              "example": "true",
              "type": "string"
            }
          },
          {
            "name": "createdAtFrom",
            "required": false,
            "in": "query",
            "description": "Ngày tạo bắt đầu để lọc",
            "schema": {
              "format": "date",
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "createdAtTo",
            "required": false,
            "in": "query",
            "description": "Ngày tạo kết thúc để lọc",
            "schema": {
              "format": "date",
              "example": "2026-08-31",
              "type": "string"
            }
          },
          {
            "name": "updatedAtFrom",
            "required": false,
            "in": "query",
            "description": "Ngày cập nhật bắt đầu để lọc",
            "schema": {
              "format": "date",
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "updatedAtTo",
            "required": false,
            "in": "query",
            "description": "Ngày cập nhật kết thúc để lọc",
            "schema": {
              "format": "date",
              "example": "2026-08-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The serialized paginated result of PostBriefResponseDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PostBriefResponseDto"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
    "/api/v1/posts/{id}": {
      "get": {
        "operationId": "PostController_getPostById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The serialized result of PostDetailResponseDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PostDetailResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      },
      "put": {
        "operationId": "PostController_updatePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostBaseUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The serialized result of OperationMetadataDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OperationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      },
      "delete": {
        "operationId": "PostController_deletePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The serialized result of OperationMetadataDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OperationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
    "/api/v1/posts/{id}/publish": {
      "put": {
        "operationId": "PostController_publishPost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The serialized result of OperationMetadataDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OperationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
    "/api/v1/posts/{id}/unpublish": {
      "put": {
        "operationId": "PostController_unpublishPost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The serialized result of OperationMetadataDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OperationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "SerializedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "description": "Dữ liệu trả về từ API"
          },
          "message": {
            "type": "string",
            "description": "Thông báo phản hồi của API",
            "example": "Thành công"
          },
          "statusCode": {
            "type": "number",
            "description": "Mã trạng thái HTTP của phản hồi",
            "example": 200
          },
          "timestamp": {
            "type": "string",
            "description": "Thời gian phản hồi",
            "example": "2023-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "data",
          "message",
          "statusCode",
          "timestamp"
        ]
      },
...
      "PostBriefResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID duy nhất của bài viết",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Tiêu đề của bài viết",
            "example": "Hướng dẫn xây dựng API NestJS chuyên nghiệp",
            "maxLength": 255
          },
          "slug": {
            "type": "string",
            "description": "Slug dùng để tạo URL thân thiện cho bài viết",
            "example": "huong-dan-xay-dung-api-nestjs-chuyen-nghiep",
            "maxLength": 255
          },
          "summary": {
            "type": "string",
            "description": "Tóm tắt ngắn gọn của bài viết",
            "example": "Giới thiệu cách tổ chức project NestJS hiệu quả",
            "maxLength": 500
          },
          "thumbnail": {
            "type": "string",
            "description": "URL ảnh thumbnail của bài viết",
            "example": "https://images.example.com/post-thumbnail.jpg",
            "format": "url"
          },
          "authorId": {
            "type": "string",
            "description": "ID của tác giả viết bài",
            "example": "d6e9f5d8-6be6-4bfe-8f13-29de6f4f2c11",
            "format": "uuid"
          },
          "published": {
            "type": "boolean",
            "description": "Trạng thái xuất bản của bài viết",
            "example": true,
            "default": false
          },
          "updatedAt": {
            "type": "string",
            "description": "Thời gian cập nhật gần nhất của bài viết",
            "example": "2026-08-20T10:15:00.000Z",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "description": "Thời gian tạo bài viết",
            "example": "2026-08-20T09:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "authorId",
          "updatedAt",
          "createdAt"
        ]
      },
      "PaginationMetadataDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Tổng số bản ghi có thể trả về",
            "example": 120,
            "minimum": 0
          },
          "totalPages": {
            "type": "number",
            "description": "Tổng số trang",
            "example": 12,
            "minimum": 0
          },
          "page": {
            "type": "number",
            "description": "Trang hiện tại",
            "example": 1,
            "minimum": 1
          },
          "limit": {
            "type": "number",
            "description": "Số lượng bản ghi trên mỗi trang",
            "example": 10,
            "minimum": 1
          }
        },
        "required": [
          "total",
          "totalPages",
          "page",
          "limit"
        ]
      },
      "PostDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID duy nhất của bài viết",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Tiêu đề của bài viết",
            "example": "Hướng dẫn xây dựng API NestJS chuyên nghiệp",
            "maxLength": 255
          },
          "slug": {
            "type": "string",
            "description": "Slug dùng để tạo URL thân thiện cho bài viết",
            "example": "huong-dan-xay-dung-api-nestjs-chuyen-nghiep",
            "maxLength": 255
          },
          "summary": {
            "type": "string",
            "description": "Tóm tắt ngắn gọn của bài viết",
            "example": "Giới thiệu cách tổ chức project NestJS hiệu quả",
            "maxLength": 500
          },
          "thumbnail": {
            "type": "string",
            "description": "URL ảnh thumbnail của bài viết",
            "example": "https://images.example.com/post-thumbnail.jpg",
            "format": "url"
          },
          "authorId": {
            "type": "string",
            "description": "ID của tác giả viết bài",
            "example": "d6e9f5d8-6be6-4bfe-8f13-29de6f4f2c11",
            "format": "uuid"
          },
          "published": {
            "type": "boolean",
            "description": "Trạng thái xuất bản của bài viết",
            "example": true,
            "default": false
          },
          "updatedAt": {
            "type": "string",
            "description": "Thời gian cập nhật gần nhất của bài viết",
            "example": "2026-08-20T10:15:00.000Z",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "description": "Thời gian tạo bài viết",
            "example": "2026-08-20T09:00:00.000Z",
            "format": "date-time"
          },
          "content": {
            "type": "string",
            "description": "Nội dung chi tiết của bài viết",
            "example": "Bài viết này mô tả cách triển khai API NestJS theo chuẩn…"
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "authorId",
          "updatedAt",
          "createdAt",
          "content"
        ]
      },
      "OperationMetadataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID duy nhất của entity",
            "example": "op_550e8400-e29b-41d4-a716-446655440000"
          },
          "success": {
            "type": "boolean",
            "description": "Kết quả thành công của thao tác",
            "example": true
          }
        },
        "required": [
          "id",
          "success"
        ]
      },
      "PostBaseCreateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Tiêu đề của bài viết",
            "example": "Hướng dẫn xây dựng API NestJS chuyên nghiệp",
            "maxLength": 255
          },
          "slug": {
            "type": "string",
            "description": "Slug dùng để tạo URL thân thiện cho bài viết",
            "example": "huong-dan-xay-dung-api-nestjs-chuyen-nghiep",
            "maxLength": 255
          },
          "content": {
            "type": "string",
            "description": "Nội dung chi tiết của bài viết",
            "example": "Bài viết này giải thích cách tổ chức API NestJS…"
          },
          "summary": {
            "type": "string",
            "description": "Tóm tắt ngắn gọn của bài viết",
            "example": "Giới thiệu cách tổ chức project NestJS hiệu quả",
            "maxLength": 500
          },
          "thumbnail": {
            "type": "string",
            "description": "URL ảnh thumbnail của bài viết",
            "example": "https://images.example.com/post-thumbnail.jpg",
            "format": "url"
          },
          "categoryId": {
            "type": "string",
            "description": "ID danh mục của bài viết",
            "example": "b733b08a-46bd-4e85-bd60-ff9e789417c2",
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "description": "ID của tác giả viết bài",
            "example": "d6e9f5d8-6be6-4bfe-8f13-29de6f4f2c11",
            "format": "uuid"
          },
          "published": {
            "type": "boolean",
            "description": "Trạng thái xuất bản của bài viết",
            "example": true,
            "default": false
          }
        },
        "required": [
          "title",
          "slug",
          "content"
        ]
      },
      "PostBaseUpdateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Tiêu đề của bài viết",
            "example": "Hướng dẫn xây dựng API NestJS chuyên nghiệp",
            "maxLength": 255
          },
          "slug": {
            "type": "string",
            "description": "Slug dùng để tạo URL thân thiện cho bài viết",
            "example": "huong-dan-xay-dung-api-nestjs-chuyen-nghiep",
            "maxLength": 255
          },
          "content": {
            "type": "string",
            "description": "Nội dung chi tiết của bài viết",
            "example": "Bài viết này giải thích cách tổ chức API NestJS…"
          },
          "summary": {
            "type": "string",
            "description": "Tóm tắt ngắn gọn của bài viết",
            "example": "Giới thiệu cách tổ chức project NestJS hiệu quả",
            "maxLength": 500
          },
          "thumbnail": {
            "type": "string",
            "description": "URL ảnh thumbnail của bài viết",
            "example": "https://images.example.com/post-thumbnail.jpg",
            "format": "url"
          },
          "categoryId": {
            "type": "string",
            "description": "ID danh mục của bài viết",
            "example": "b733b08a-46bd-4e85-bd60-ff9e789417c2",
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "description": "ID của tác giả viết bài",
            "example": "d6e9f5d8-6be6-4bfe-8f13-29de6f4f2c11",
            "format": "uuid"
          },
          "published": {
            "type": "boolean",
            "description": "Trạng thái xuất bản của bài viết",
            "example": true,
            "default": false
          },
          "publishedAt": {
            "type": "object",
            "description": "Thời gian bài viết được xuất bản",
            "example": "2026-08-20T10:00:00.000Z",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "apiKeyCookie",
        "description": "API Key authentication via cookie"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication"
      }
    }
  }
}